+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/NavigationOptions.java
@@ -6,19 +6,19 @@ import de.mephisto.vpin.connectors.vps.model.VpsTable;
  * Contains parameters that can be used to initialize a navigation item.
  */
 public class NavigationOptions {
-  private VpsTable vpsTable;
+  private Object model;
   private int gameId;
 
   public NavigationOptions(int gameId) {
     this.gameId = gameId;
   }
 
-  public NavigationOptions(VpsTable vpsTable) {
-    this.vpsTable = vpsTable;
+  public NavigationOptions(Object model) {
+    this.model = model;
   }
 
-  public VpsTable getVpsTable() {
-    return vpsTable;
+  public Object getModel() {
+    return model;
   }
 
   public int getGameId() {