+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/mania/TabManiaPlayerStatsController.java
@@ -30,7 +30,7 @@ public class TabManiaPlayerStatsController implements Initializable, StudioFXCon
 
   @Override
   public void onViewActivated(@Nullable NavigationOptions options) {
-    playerStatsWidgetController.onViewActivated();
+    playerStatsWidgetController.onViewActivated(options);
   }
 
   @Override