+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TableOverviewColumnSorter.java
@@ -103,6 +103,12 @@ public class TableOverviewColumnSorter {
           comp = comp.reversed();
         }
       }
+      else if (column.equals(tableOverviewController.columnRES)) {
+        comp = Comparator.comparing(o -> o.getGame().getResPath() != null);
+        if (column.getSortType().equals(TableColumn.SortType.DESCENDING)) {
+          comp = comp.reversed();
+        }
+      }
       else if (column.equals(tableOverviewController.columnINI)) {
         comp = Comparator.comparing(o -> o.getGame().getIniPath() != null);
         if (column.getSortType().equals(TableColumn.SortType.DESCENDING)) {