+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/util/AutoCompleteMatcher.java
@@ -1,0 +1,8 @@
+package de.mephisto.vpin.ui.util;
+
+import java.util.List;
+
+public interface AutoCompleteMatcher {
+
+  List<String> match(String input);
+}